game-api/requirements.txt
Flask
psycopg2-binary
pgvector
google-generativeai
